requirements.txt
openai
langchain
streamlit==1.8.0
python-dotenv 
PyPDF2
langchain-community
altair==4.2.0
langchain-openai






-e .